iommu/arm-smmu-v3: Add in-kernel support for NVIDIA Tegra241 (Grace) CMDQ-V
NVIDIA's Tegra241 Soc has a CMDQ-Virtualization (CMDQV) hardware, extending
the standard ARM SMMU v3 IP to support multiple V-CMDQs with virtualization
capabilities. In terms of command queue, they are very like a standard SMMU
CMDQ (or ECMDQs), but only support CS_NONE in the CS field of CMD_SYNC.

Add a new tegra241-cmdqv driver, and insert its structure pointer into the
existing arm_smmu_device, and then add related function calls in the SMMUv3
driver.

In the CMDQV driver itself, add minimal part for an in-kernel support. This
allows OS (either a host or a guest kernel) to detect if it's running on an
NVIDIA SoC with VCMDQ, in which case: reserve VINTF0 for in-kernel use, and
assign some of the VCMDQs to the VINTF0, and select one of the VCMDQs based
on the currently running CPU ID to issue commands.

Note that for now the CMDQV driver only supports an ACPI configuration. And
for in-kernel support of a guest kernel. The host kernel must provide VCMDQ
user-space support that will come with an additional series of patches.

Signed-off-by: Nate Watterson <nwatterson@nvidia.com>
Co-developed-by: Nicolin Chen <nicolinc@nvidia.com>
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>