vfio/nvgrace-gpu: Add vfio pci variant module for grace hopper
NVIDIA's upcoming Grace Hopper Superchip provides a PCI-like device
for the on-chip GPU that is the logical OS representation of the
internal proprietary chip-to-chip cache coherent interconnect.

The device is peculiar compared to a real PCI device in that whilst
there is a real 64b PCI BAR1 (comprising region 2 & region 3) on the device,
it is not used to access device memory once the faster chip-to-chip
interconnect is initialized (occurs at the time of host system boot).
The device memory is accessed instead using the chip-to-chip interconnect
that is exposed as a contiguous physically addressable region on the host.

Provide a VFIO PCI variant driver that adapts the unique device memory
representation into a more standard PCI representation facing userspace.

The device memory aperture is obtained from ACPI using
device_property_read_u64(), according to the FW specification, and exported
to userspace as a separate VFIO BAR region. Since the device implements
64-bit BAR0, the VFIO PCI variant driver maps the GPU memory aperture
to the next available PCI BAR (BAR2). Qemu will then naturally generate a
PCI device in the VM with the cacheable aperture reported in BAR2 region. The
variant driver also provides emulation for this fake BAR's PCI config space
offset registers.

Since this memory region is actually cache coherent with the CPU, the
VFIO variant driver will mmap it into VMA using a cacheable mapping. The
mapping is done using remap_pfn_range().

PCI BAR are aligned to the power-of-2, but the actual memory on the
device may not. A read or write access to the physical address from the
last device PFN up to the next power-of-2 aligned physical address
results in reading ~0 and dropped writes. Note that the GPU device
driver [1] is capable of knowing the exact device memory size through
separate means. The device memory size is primarily kept in the system
ACPI tables for use by the VFIO PCI variant module.

This goes along with a qemu series to provides the necessary
implementation of the Grace Hopper Superchip firmware specification so
that the guest operating system can see the correct ACPI modeling for
the coherent GPU device. Verified with the CUDA workload in the VM.
https://lore.kernel.org/all/20231107190039.19434-1-ankita@nvidia.com/

This patch is split from a patch series being pursued separately:
https://lore.kernel.org/lkml/20230405180134.16932-1-ankita@nvidia.com/

Applied over next-20231113.

[1] https://github.com/NVIDIA/open-gpu-kernel-modules